backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pillow==10.4.0
numpy==2.1.1
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
